PIM-3832: introduce ProductFieldUpdater interface (extracted from ProductUpdaterInterface)
