+++ spec/Pim/Bundle/CatalogBundle/Updater/ProductTemplateUpdaterSpec.php
@@ -5,13 +5,13 @@ namespace spec\Pim\Bundle\CatalogBundle\Updater;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
-use Pim\Bundle\CatalogBundle\Updater\ProductUpdaterInterface;
+use Pim\Bundle\CatalogBundle\Updater\ProductFieldUpdaterInterface;
 
 class ProductTemplateUpdaterSpec extends ObjectBehavior
 {
-    function let(ProductUpdaterInterface $productUpdater)
+    function let(ProductFieldUpdaterInterface $productFieldUpdater)
     {
-        $this->beConstructedWith($productUpdater);
+        $this->beConstructedWith($productFieldUpdater);
     }
 
     function it_is_initializable()
@@ -25,7 +25,7 @@ class ProductTemplateUpdaterSpec extends ObjectBehavior
     }
 
     function it_updates_products_with_variant_group_template_values_using_product_updater(
-        $productUpdater,
+        $productFieldUpdater,
         ProductTemplateInterface $template,
         ProductInterface $product
     ) {
@@ -73,10 +73,10 @@ class ProductTemplateUpdaterSpec extends ObjectBehavior
 
         $template->getValuesData()->willReturn($updates);
 
-        $productUpdater->setData($product, 'description', 'Foo', ['locale' => 'en_US', 'scope' => 'ecommerce'])->shouldBeCalled();
-        $productUpdater->setData($product, 'description', 'Bar', ['locale' => 'en_US', 'scope' => 'mobile'])->shouldBeCalled();
-        $productUpdater->setData($product, 'color', 'red', ['locale' => null, 'scope' => null])->shouldBeCalled();
-        $productUpdater
+        $productFieldUpdater->setData($product, 'description', 'Foo', ['locale' => 'en_US', 'scope' => 'ecommerce'])->shouldBeCalled();
+        $productFieldUpdater->setData($product, 'description', 'Bar', ['locale' => 'en_US', 'scope' => 'mobile'])->shouldBeCalled();
+        $productFieldUpdater->setData($product, 'color', 'red', ['locale' => null, 'scope' => null])->shouldBeCalled();
+        $productFieldUpdater
             ->setData(
                 $product,
                 'price',
@@ -84,7 +84,7 @@ class ProductTemplateUpdaterSpec extends ObjectBehavior
                 ['locale' => 'fr_FR', 'scope' => null]
             )
             ->shouldBeCalled();
-        $productUpdater
+        $productFieldUpdater
             ->setData(
                 $product,
                 'image',
