+++ spec/Pim/Bundle/EnrichBundle/Processor/MassEdit/AddProductValueProcessorSpec.php
@@ -5,10 +5,10 @@ namespace spec\Pim\Bundle\EnrichBundle\Processor\MassEdit;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use PhpSpec\ObjectBehavior;
+use Pim\Bundle\BaseConnectorBundle\Model\JobConfigurationInterface;
+use Pim\Bundle\BaseConnectorBundle\Model\Repository\JobConfigurationRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
-use Pim\Bundle\CatalogBundle\Updater\ProductUpdaterInterface;
-use Pim\Component\Connector\Model\JobConfigurationInterface;
-use Pim\Component\Connector\Repository\JobConfigurationRepositoryInterface;
+use Pim\Bundle\CatalogBundle\Updater\ProductFieldUpdaterInterface;
 use Prophecy\Argument;
 use Symfony\Component\Validator\ConstraintViolation;
 use Symfony\Component\Validator\ConstraintViolationList;
@@ -17,19 +17,19 @@ use Symfony\Component\Validator\ValidatorInterface;
 class AddProductValueProcessorSpec extends ObjectBehavior
 {
     function let(
-        ProductUpdaterInterface $productUpdater,
+        ProductFieldUpdaterInterface $productFieldUpdater,
         ValidatorInterface $validator,
         JobConfigurationRepositoryInterface $jobConfigurationRepo
     ) {
         $this->beConstructedWith(
-            $productUpdater,
+            $productFieldUpdater,
             $validator,
             $jobConfigurationRepo
         );
     }
 
     function it_adds_values_to_product(
-        $productUpdater,
+        $productFieldUpdater,
         $validator,
         ProductInterface $product,
         StepExecution $stepExecution,
@@ -47,7 +47,7 @@ class AddProductValueProcessorSpec extends ObjectBehavior
             json_encode(['filters' => [], 'actions' => [['field' => 'categories', 'value' => ['office', 'bedroom']]]])
         );
 
-        $productUpdater->addData($product, 'categories', ['office', 'bedroom'])->shouldBeCalled();
+        $productFieldUpdater->addData($product, 'categories', ['office', 'bedroom'])->shouldBeCalled();
         $stepExecution->incrementSummaryInfo('mass_edited')->shouldBeCalled();
 
         $this->setStepExecution($stepExecution);
@@ -56,7 +56,7 @@ class AddProductValueProcessorSpec extends ObjectBehavior
     }
 
     function it_adds_invalid_values_to_product(
-        $productUpdater,
+        $productFieldUpdater,
         $validator,
         ProductInterface $product,
         StepExecution $stepExecution,
@@ -75,7 +75,7 @@ class AddProductValueProcessorSpec extends ObjectBehavior
             json_encode(['filters' => [], 'actions' => [['field' => 'categories', 'value' => ['office', 'bedroom']]]])
         );
 
-        $productUpdater->addData($product, 'categories', ['office', 'bedroom'])->shouldBeCalled();
+        $productFieldUpdater->addData($product, 'categories', ['office', 'bedroom'])->shouldBeCalled();
         $stepExecution->addWarning(Argument::cetera())->shouldBeCalled();
         $stepExecution->incrementSummaryInfo('skipped_products')->shouldBeCalled();
