+++ src/Pim/Bundle/CatalogBundle/Resources/config/updaters.yml
@@ -1,5 +1,6 @@
 parameters:
     pim_catalog.updater.product.class:                       Pim\Bundle\CatalogBundle\Updater\ProductUpdater
+    pim_catalog.updater.product_field.class:                 Pim\Bundle\CatalogBundle\Updater\ProductFieldUpdater
     pim_catalog.updater.product_template.class:              Pim\Bundle\CatalogBundle\Updater\ProductTemplateUpdater
     pim_catalog.updater.attribute_option.class:              Pim\Bundle\CatalogBundle\Updater\AttributeOptionUpdater
 
@@ -54,6 +55,11 @@ services:
     pim_catalog.updater.product:
         class: %pim_catalog.updater.product.class%
         arguments:
+            - @pim_catalog.updater.product_field
+
+    pim_catalog.updater.product_field:
+        class: %pim_catalog.updater.product_field.class%
+        arguments:
             - @pim_catalog.repository.attribute
             - @pim_catalog.updater.setter.registry
             - @pim_catalog.updater.copier.registry
@@ -63,7 +69,7 @@ services:
     pim_catalog.updater.product_template:
         class: %pim_catalog.updater.product_template.class%
         arguments:
-            - @pim_catalog.updater.product
+            - @pim_catalog.updater.product_field
 
     pim_catalog.updater.attribute_option:
         class: %pim_catalog.updater.attribute_option.class%
