+++ src/Pim/Bundle/CatalogBundle/Updater/ProductUpdater.php
@@ -2,13 +2,8 @@
 
 namespace Pim\Bundle\CatalogBundle\Updater;
 
-use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
+use Doctrine\Common\Util\ClassUtils;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
-use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Updater\Adder\AdderRegistryInterface;
-use Pim\Bundle\CatalogBundle\Updater\Copier\CopierRegistryInterface;
-use Pim\Bundle\CatalogBundle\Updater\Remover\RemoverRegistryInterface;
-use Pim\Bundle\CatalogBundle\Updater\Setter\SetterRegistryInterface;
 
 /**
  * Provides basic operations to update a product
@@ -19,50 +14,35 @@ use Pim\Bundle\CatalogBundle\Updater\Setter\SetterRegistryInterface;
  */
 class ProductUpdater implements ProductUpdaterInterface
 {
-    /** @var AttributeRepositoryInterface */
-    protected $attributeRepository;
-
-    /** @var SetterRegistryInterface */
-    protected $setterRegistry;
-
-    /** @var CopierRegistryInterface */
-    protected $copierRegistry;
-
-    /** @var AdderRegistryInterface */
-    protected $adderRegistry;
-
-    /** @var RemoverRegistryInterface */
-    protected $removerRegistry;
+    /** @var ProductFieldUpdaterInterface */
+    protected $productFieldUpdater;
 
     /**
-     * @param AttributeRepositoryInterface $repository
-     * @param SetterRegistryInterface      $setterRegistry
-     * @param CopierRegistryInterface      $copierRegistry
-     * @param AdderRegistryInterface       $adderRegistry
+     * @param ProductFieldUpdaterInterface $productFieldUpdater
      */
-    public function __construct(
-        AttributeRepositoryInterface $repository,
-        SetterRegistryInterface $setterRegistry,
-        CopierRegistryInterface $copierRegistry,
-        AdderRegistryInterface $adderRegistry,
-        RemoverRegistryInterface $removerRegistry
-    ) {
-        $this->attributeRepository = $repository;
-        $this->setterRegistry      = $setterRegistry;
-        $this->copierRegistry      = $copierRegistry;
-        $this->adderRegistry       = $adderRegistry;
-        $this->removerRegistry     = $removerRegistry;
+    public function __construct(ProductFieldUpdaterInterface $productFieldUpdater)
+    {
+        $this->productFieldUpdater = $productFieldUpdater;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function update(ProductInterface $product, array $data, array $options = [])
+    public function update($product, array $data, array $options = [])
     {
+        if (!$product instanceof ProductInterface) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Expects a "Pim\Bundle\CatalogBundle\Model\ProductInterface", "%s" provided.',
+                    ClassUtils::getClass($product)
+                )
+            );
+        }
+
         foreach ($data as $field => $values) {
             // TODO: hard coded :(
             if (in_array($field, ['enabled', 'categories', 'groups', 'associations'])) {
-                $this->setData($product, $field, $values, []);
+                $this->productFieldUpdater->setData($product, $field, $values, []);
             } else {
                 foreach ($values as $value) {
                     // sets the value if the attribute belongs to the family or if the value already exists as optional
@@ -71,7 +51,7 @@ class ProductUpdater implements ProductUpdaterInterface
                     $hasValue = $product->getValue($field, $value['locale'], $value['scope']) !== null;
                     if ($belongsToFamily || $hasValue) {
                         $options = ['locale' => $value['locale'], 'scope' => $value['scope']];
-                        $this->setData($product, $field, $value['data'], $options);
+                        $this->productFieldUpdater->setData($product, $field, $value['data'], $options);
                     }
                 }
             }
@@ -81,116 +61,10 @@ class ProductUpdater implements ProductUpdaterInterface
     /**
      * {@inheritdoc}
      */
-    public function setData(ProductInterface $product, $field, $data, array $options = [])
-    {
-        $attribute = $this->getAttribute($field);
-        if (null !== $attribute) {
-            $setter = $this->setterRegistry->getAttributeSetter($attribute);
-        } else {
-            $setter = $this->setterRegistry->getFieldSetter($field);
-        }
-
-        if (null === $setter) {
-            throw new \LogicException(sprintf('No setter found for field "%s"', $field));
-        }
-
-        if (null !== $attribute) {
-            $setter->setAttributeData($product, $attribute, $data, $options);
-        } else {
-            $setter->setFieldData($product, $field, $data, $options);
-        }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addData(ProductInterface $product, $field, $data, array $options = [])
-    {
-        $attribute = $this->getAttribute($field);
-        if (null !== $attribute) {
-            $adder = $this->adderRegistry->getAttributeAdder($attribute);
-        } else {
-            $adder = $this->adderRegistry->getFieldAdder($field);
-        }
-
-        if (null === $adder) {
-            throw new \LogicException(sprintf('No adder found for field "%s"', $field));
-        }
-
-        if (null !== $attribute) {
-            $adder->addAttributeData($product, $attribute, $data, $options);
-        } else {
-            $adder->addFieldData($product, $field, $data, $options);
-        }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function copyData(
-        ProductInterface $fromProduct,
-        ProductInterface $toProduct,
-        $fromField,
-        $toField,
-        array $options = []
-    ) {
-        $fromAttribute = $this->getAttribute($fromField);
-        $toAttribute = $this->getAttribute($toField);
-        if (null !== $fromAttribute && null !== $toAttribute) {
-            $copier = $this->copierRegistry->getAttributeCopier($fromAttribute, $toAttribute);
-        } else {
-            $copier = $this->copierRegistry->getFieldCopier($fromField, $toField);
-        }
-
-        if (null === $copier) {
-            throw new \LogicException(sprintf('No copier found for fields "%s" and "%s"', $fromField, $toField));
-        }
-
-        if (null !== $fromAttribute && null !== $toAttribute) {
-            $copier->copyAttributeData($fromProduct, $toProduct, $fromAttribute, $toAttribute, $options);
-        } else {
-            $copier->copyFieldData($fromProduct, $toProduct, $fromField, $toField, $options);
-        }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function removeData(ProductInterface $product, $field, $data, array $options = [])
-    {
-        $attribute = $this->getAttribute($field);
-        if (null !== $attribute) {
-            $remover = $this->removerRegistry->getAttributeRemover($attribute);
-        } else {
-            $remover = $this->removerRegistry->getFieldRemover($field);
-        }
-
-        if (null === $remover) {
-            throw new \LogicException(sprintf('No remover found for field "%s"', $field));
-        }
-
-        if (null !== $attribute) {
-            $remover->removeAttributeData($product, $attribute, $data, $options);
-        } else {
-            $remover->removeFieldData($product, $field, $data, $options);
-        }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function setValue(array $products, $field, $data, $locale = null, $scope = null)
     {
         foreach ($products as $product) {
-            $this->setData($product, $field, $data, ['locale' => $locale, 'scope' => $scope]);
+            $this->productFieldUpdater->setData($product, $field, $data, ['locale' => $locale, 'scope' => $scope]);
         }
 
         return $this;
@@ -215,21 +89,9 @@ class ProductUpdater implements ProductUpdaterInterface
             'to_scope' => $toScope,
         ];
         foreach ($products as $product) {
-            $this->copyData($product, $product, $fromField, $toField, $options);
+            $this->productFieldUpdater->copyData($product, $product, $fromField, $toField, $options);
         }
 
         return $this;
-    }
-
-    /**
-     * @param string $code
-     *
-     * @return AttributeInterface|null
-     */
-    protected function getAttribute($code)
-    {
-        $attribute = $this->attributeRepository->findOneBy(['code' => $code]);
-
-        return $attribute;
     }
 }
