+++ src/Pim/Bundle/EnrichBundle/Resources/config/processors.yml
@@ -10,21 +10,21 @@ services:
     pim_enrich.mass_edit.update_product_value.processor:
         class: %pim_enrich.mass_edit.update_product_value.processor.class%
         arguments:
-            - '@pim_catalog.updater.product'
+            - '@pim_catalog.updater.product_field'
             - '@pim_validator'
             - '@pim_base_connector.repository.job_configuration'
 
     pim_enrich.mass_edit.add_product_value.processor:
         class: %pim_enrich.mass_edit.add_product_value.processor.class%
         arguments:
-            - '@pim_catalog.updater.product'
+            - '@pim_catalog.updater.product_field'
             - '@pim_validator'
             - '@pim_base_connector.repository.job_configuration'
 
     pim_enrich.mass_edit.edit_common_attributes.processor:
         class: %pim_enrich.mass_edit.edit_common_attributes.processor.class%
         arguments:
-            - '@pim_catalog.updater.product'
+            - '@pim_catalog.updater.product_field'
             - '@pim_validator'
             - '@pim_catalog.repository.product_mass_action'
             - '@pim_catalog.repository.attribute'
